server/requirements.txt
Flask==2.3.3
requests==2.31.0
PyMuPDF==1.23.7
google-generativeai==0.3.2
